minor update 3d tool (spf) and projmanager, ipestp adjustments to save intermediate results
git-svn-id: https://svn.oss.deltares.nl/repos/imod/trunk@2992 8a8853d3-af7d-47ec-b359-4bd2096ba547

+++ MODFLOW/iMODUTL/imod_pest_mod.f90
@@ -60,8 +60,8 @@ TYPE PARAOBJ
  LOGICAL :: LOG             !## logical to determine whether parameter is lognormal
  INTEGER(KIND=2),POINTER,DIMENSION(:) :: IROW,ICOL  !## node number of zone for parameter param(i) --- pointer to IZONE
  REAL(KIND=8),POINTER,DIMENSION(:,:) :: XY  !## xy location for pilot-points
- REAL(KIND=8),POINTER,DIMENSION(:) :: X  !## copy of variable to be adjusted
- REAL(KIND=8),POINTER,DIMENSION(:) :: F  !## fraction of variable to be adjusted
+ REAL(KIND=4),POINTER,DIMENSION(:) :: X  !## copy of variable to be adjusted
+ REAL(KIND=4),POINTER,DIMENSION(:) :: F  !## fraction of variable to be adjusted
 END TYPE PARAOBJ
 TYPE(PARAOBJ),ALLOCATABLE,DIMENSION(:) :: PARAM
 TYPE ZONEOBJ
